# contracts/v0.1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "infer-runtime consumer API",
    "version": "0.1.0-candidate.2",
    "description": "Candidate contract for application integration. This revision is immutable; later candidate revisions are additive unless accompanied by migration notes. Operator resource and provider administration routes are intentionally excluded."
  },
  "servers": [
    { "url": "http://127.0.0.1:8787" }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/v1/responses": {
      "post": {
        "operationId": "createResponse",
        "description": "Submit the supported stateless subset of an OpenAI Responses request. The model field is an infer-runtime intent, not a physical deployment ID.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResponsesRequest" },
              "examples": {
                "minimal": { "$ref": "#/components/examples/ResponsesRequest" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Completed response, or an SSE stream when stream=true",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ResponseObject" }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE event/data frames. Provider event types are forwarded after response identity normalization; runtime-generated failures use event: error."
                }
              }
            }
          },
          "202": {
            "description": "Durable local background job accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BackgroundResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/responses/{response_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ResponseId" }
      ],
      "get": {
        "operationId": "getResponse",
        "description": "Get a durable background response owned by the authenticated application.",
        "responses": {
          "200": {
            "description": "Current or completed background response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BackgroundResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/infer/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } },
          { "name": "cursor", "in": "query", "description": "Opaque continuation cursor", "schema": { "type": "string" } },
          { "name": "priority", "in": "query", "schema": { "$ref": "#/components/schemas/Priority" } },
          { "name": "state", "in": "query", "schema": { "$ref": "#/components/schemas/JobState" } }
        ],
        "responses": {
          "200": {
            "description": "Application-scoped job page",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobListPage" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/infer/v1/jobs/{response_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ResponseId" }
      ],
      "get": {
        "operationId": "getJob",
        "responses": {
          "200": {
            "description": "Full application-scoped job snapshot",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobSnapshot" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Application credential configured in infer-runtime"
      }
    },
    "parameters": {
      "ResponseId": {
        "name": "response_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "minLength": 1 }
      }
    },
...
    "responses": {
      "Error": {
        "description": "Stable error envelope; code is machine-actionable and message is diagnostic text",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "examples": {
              "invalidRequest": { "$ref": "#/components/examples/Error" }
            }
          }
        }
      }
    },
    "examples": {
      "ResponsesRequest": {
        "value": {
          "model": "text.summarize",
          "input": "Summarize this text.",
          "metadata": {
            "infer.placement": "local_only",
            "infer.priority": "normal"
          }
        }
      },
      "Error": {
        "value": {
          "error": {
            "message": "request body does not match the documented schema",
            "type": "invalid_request_error",
            "code": "invalid_request_error"
          }
        }
      }
    },
    "schemas": {
...
      "ResponsesRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["model", "input"],
        "properties": {
          "model": { "type": "string", "minLength": 1, "description": "Stable intent profile, never a physical deployment ID" },
          "input": {},
          "instructions": {},
          "stream": { "type": "boolean", "default": false },
          "background": { "type": "boolean", "default": false },
          "metadata": { "$ref": "#/components/schemas/Metadata" },
          "tools": { "type": "array", "items": {} },
          "reasoning": { "$ref": "#/components/schemas/Reasoning" },
          "temperature": { "type": ["number", "null"] },
          "top_p": { "type": ["number", "null"] },
          "max_output_tokens": { "type": ["integer", "null"], "minimum": 0, "maximum": 4294967295 },
          "truncation": { "type": ["string", "null"] },
          "store": { "enum": [false, null], "description": "Only false or null is accepted" },
          "previous_response_id": { "type": "null", "description": "Stateful chaining is not supported" },
          "conversation": { "type": "null", "description": "Conversation state is not supported" }
        }
      },
      "Metadata": {
        "type": "object",
        "default": {},
        "additionalProperties": { "type": "string" },
        "description": "Application metadata plus documented infer.* routing constraints. Unknown reserved infer.* keys are rejected."
      },
      "Reasoning": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "effort": {
            "type": ["string", "null"],
            "enum": ["none", "low", "medium", "high", "xhigh", "max", null]
          }
        }
      },
      "ResponseObject": {
        "type": "object",
        "additionalProperties": true,
        "required": ["id", "object", "created_at", "model"],
        "properties": {
          "id": { "type": "string" },
          "object": { "const": "response" },
          "created_at": { "type": "integer" },
          "model": { "type": "string", "description": "Logical intent, not physical deployment" },
          "status": { "type": "string" },
          "output": { "type": "array" },
          "parallel_tool_calls": { "type": "boolean" },
          "tool_choice": {},
          "tools": { "type": "array" }
        }
      },
      "BackgroundResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseObject" },
          {
            "type": "object",
            "required": ["background", "status", "output"],
            "properties": {
              "background": { "const": true },
              "status": { "enum": ["queued", "in_progress", "completed", "failed", "cancelled", "incomplete"] },
              "output": { "type": "array" },
              "error": { "type": ["object", "null"] }
            }
          }
        ]
      },
...
      "Priority": {
        "type": "string",
        "enum": ["interactive", "normal", "background"]
      },
      "JobState": {
        "type": "string",
        "enum": ["queued", "running", "succeeded", "failed", "cancelled", "expired"]
      },
      "JobListItem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "app_id", "intent", "provider", "deployment", "state", "priority", "created_at_ms", "updated_at_ms"],
        "properties": {
          "id": { "type": "string" },
          "app_id": { "type": "string" },
          "intent": { "type": "string" },
          "provider": { "type": "string" },
          "deployment": { "type": "string" },
          "state": { "$ref": "#/components/schemas/JobState" },
          "priority": { "$ref": "#/components/schemas/Priority" },
          "created_at_ms": { "type": "integer" },
          "updated_at_ms": { "type": "integer" }
        }
      },
      "JobListPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["jobs"],
        "properties": {
          "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/JobListItem" } },
          "next_cursor": { "type": "string", "description": "Opaque continuation cursor" }
        }
      },
      "JobSnapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "app_id", "intent", "provider", "deployment", "model_profile", "model_build", "physical_model", "placement", "quality_grade", "rating_status", "resource_class", "state", "policy", "priority", "constraints", "routing", "attempts", "error"],
        "properties": {
          "id": { "type": "string" },
          "app_id": { "type": "string" },
          "intent": { "type": "string" },
          "provider": { "type": "string" },
          "deployment": { "type": "string" },
          "model_profile": { "type": "string" },
          "model_build": { "type": "string" },
          "physical_model": { "type": "string" },
          "placement": { "enum": ["local", "trusted_node", "cloud"] },
          "quality_grade": { "enum": ["basic", "general", "advanced", "frontier"] },
          "rating_status": { "enum": ["provisional", "benchmarked"] },
          "resource_class": { "enum": ["light", "standard", "heavy", "extreme"] },
          "state": { "$ref": "#/components/schemas/JobState" },
          "policy": { "type": "string" },
          "priority": { "$ref": "#/components/schemas/Priority" },
          "constraints": { "type": "object", "additionalProperties": true },
          "routing": { "$ref": "#/components/schemas/RoutingDecision" },
          "attempts": { "type": "array", "items": { "$ref": "#/components/schemas/Attempt" } },
          "error": { "type": ["string", "null"] }
        }
      },
      "RoutingDecision": {
        "type": "object",
        "additionalProperties": false,
        "required": ["quality_floor", "candidates"],
        "properties": {
          "quality_floor": { "enum": ["basic", "general", "advanced", "frontier"] },
          "candidates": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["deployment", "provider", "status"],
              "properties": {
                "deployment": { "type": "string" },
                "provider": { "type": "string" },
                "status": { "enum": ["eligible", "fallback_eligible", "rejected"] },
                "rank": { "type": "integer", "minimum": 0 },
                "reason_codes": { "type": "array", "items": { "type": "string" } }
              }
            }
          }
        }
      },
      "Attempt": {
        "type": "object",
        "additionalProperties": false,
        "required": ["number", "provider", "deployment", "outcome", "trigger"],
        "properties": {
          "number": { "type": "integer", "minimum": 0 },
          "provider": { "type": "string" },
          "deployment": { "type": "string" },
          "outcome": { "enum": ["running", "succeeded", "failed", "interrupted"] },
          "trigger": { "enum": ["initial", "retry", "fallback", "recovery"] },
          "error_kind": { "type": "string" },
          "error": { "type": "string" }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["message", "type", "code"],
            "properties": {
              "message": { "type": "string", "description": "Diagnostic text; not a stable machine discriminator" },
              "type": { "const": "invalid_request_error" },
              "code": { "type": "string", "description": "Stable machine-actionable error code" }
            }
          }
        }
      }
    }
  }
}